Crate rtf_integrations

The integrations with external APIs for the Apollo Runtime Testing Framework.

Modules

github

A lightweight GitHub API client

graphos

Logic for working with Apollo GraphOS

Structs

HttpResponse

Represents the result of an HTTP request made by an HttpClient.

ReqwestClient

A client implementation that is backed by a [reqwest::Client].

Constants

APOLLO_KEY_ENV_VAR

The environment variable name for the api key used to authenticate with the GraphOS API

APOLLO_SUDO_ENV_VAR

The environment variable name for setting the apollo-sudo=true header in GraphOS API requests

GITHUB_TOKEN_ENV_VAR

The environment variable name for the api token used to authenticate with the GitHub API.

GRAPH_OS_STAGING_ENV_VAR

The environment variable name for whether or not to use the staging GraphOS API

N_PARALLEL_FETCH

The maximum number of queries to run in parallel querying the platform API.

Traits

HttpClient

Types that implement HttpClient may be used to perform http requests