rtf_integrations::github

Trait Client

pub trait Client {
    // Required method
    async fn raw_file_content(
        &self,
        org: &str,
        repo: &str,
        path: &str,
        git_ref: Option<impl AsRef<str>>,
    ) -> Result<Bytes, Error>;

    // Provided method
    async fn string_file_content(
        &self,
        org: &str,
        repo: &str,
        path: &str,
        git_ref: Option<impl AsRef<str>>,
    ) -> Result<String, Error> { ... }
}

An API client that can make requests to the GitHub REST API.

Required Methods

async fn raw_file_content( &self, org: &str, repo: &str, path: &str, git_ref: Option<impl AsRef<str>>, ) -> Result<Bytes, Error>

Attempt to pull the raw file content of a given file as [Bytes] from the specified GitHub repo.

The API token used to create this client must have access to the repo in question.

Provided Methods

async fn string_file_content( &self, org: &str, repo: &str, path: &str, git_ref: Option<impl AsRef<str>>, ) -> Result<String, Error>

Attempt to pull the raw file content of a given file as a utf-8 String from the specified GitHub repo.

The API token used to create this client must have access to the repo in question.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.