rtf_integrations::github

Trait Client

pub trait Client: Send + Sync {
    // Required method
    fn raw_file_content<G: AsRef<str> + Send>(
        &self,
        org: &str,
        repo: &str,
        path: &str,
        git_ref: Option<G>,
    ) -> impl Future<Output = Result<Bytes, Error>> + Send;

    // Provided method
    fn string_file_content<G: AsRef<str> + Send>(
        &self,
        org: &str,
        repo: &str,
        path: &str,
        git_ref: Option<G>,
    ) -> impl Future<Output = Result<String, Error>> + Send { ... }
}

An API client that can make requests to the GitHub REST API.

Required Methods

fn raw_file_content<G: AsRef<str> + Send>( &self, org: &str, repo: &str, path: &str, git_ref: Option<G>, ) -> impl Future<Output = Result<Bytes, Error>> + Send

Attempt to pull the raw file content of a given file as [Bytes] from the specified GitHub repo.

The API token used to create this client must have access to the repo in question.

Provided Methods

fn string_file_content<G: AsRef<str> + Send>( &self, org: &str, repo: &str, path: &str, git_ref: Option<G>, ) -> impl Future<Output = Result<String, Error>> + Send

Attempt to pull the raw file content of a given file as a utf-8 String from the specified GitHub repo.

The API token used to create this client must have access to the repo in question.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.